+++ changelog.md
@@ -4,6 +4,7 @@
 - Re-Added packing and unpacking recipes for the metal press (BluSunrize)
 - Added cloche recipe to grow bamboo (BluSunrize)
 - Added composting for hempseeds (BluSunrize)
+- Added missing recipe for the "Common Projectiles" blueprint (BluSunrize)
 - Fixed bounding boxes for vertical covered conveyors (Malte)
 - Fixed metal press not forming when mirrored (Malte)
 - Fixed extracting conveyors not saving their extract direction (BluSunrize)